requirements.txt
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.6
Flask==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.7.0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
SecretStorage==3.3.3
urllib3==2.5.0
Werkzeug==3.1.3
